+++ Changelog.md
@@ -19,3 +19,4 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Initial version
 - Connect to MQTT broker
 - Wait for message and display
+- Terminate on "exit" message